Pengembangan Website::saveLandingPageImage: Pengecekan apakah oldImage hero adalah null atau tidak

+++ app/Controllers/admin/Website.php
@@ -163,8 +163,10 @@
                 if(!empty($oldHeroImage)){
                     $oldImage       =   $oldHeroImage['value'];
                     $oldImage       =   uploadGambarWebsite('landing-page/'.$oldImage);
-                    if(file_exists($oldImage)){
-                        unlink($oldImage);
+                    if(!empty($oldImage)){
+                        if(file_exists($oldImage)){
+                            unlink($oldImage);
+                        }
                     }
                 }
